requirements.txt
# h5py==3.11.0
# librosa==0.10.2.post1
# matplotlib==3.10.8
# numpy~=1.26.4
# opensmile==2.5.0
# pandas==3.0.2
# pydantic==2.13.0
# pydub==0.25.1
# PyYAML==6.0.1
# Requests==2.33.1
# scikit_learn==1.4.2
# scipy==1.17.1
# skrebate==0.62
# soundfile==0.12.1
# spafe==0.3.3
# speechbrain==1.0.1
# tensorflow==2.14.0
# tensorflow_hub==0.15.0
# # torch==2.4.1+cu118
# # torchaudio==2.4.1+cu118
# tqdm==4.66.4
# transformers==4.48.3
# umap_learn==0.5.7
# xgboost==3.2.0



pandas~=2.2.1
scikit-learn~=1.4.2
numpy~=1.26.4
tqdm~=4.66.2
librosa~=0.10.2
soundfile~=0.12.1
faker~=24.14.1
opensmile~=2.5.0
spafe~=0.3.2
requests~=2.32.0
pydub~=0.25.1
torch==2.4.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.4.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.19.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
transformers==4.37.2
ffmpeg-python>=0.2
setuptools>=65
datasets~=3.0.0
pyannote.audio~=3.3.2
pyyaml~=6.0.1
tensorflow==2.14.0
tensorflow_hub==0.15.0
tensorboard==2.14.1
tensorflow-estimator==2.14.0
xgboost==2.1.1
pydantic==2.5.3
stanza==1.10.1
gensim==4.3.3
nltk==3.9.1
tables==3.10.2